Refactor Header/index to more-minimal implementation; links only used in HeaderSmall
