+++ src/components/Header/HeaderSmall.js
@@ -6,7 +6,25 @@ import DropdownList from './DropdownList';
 import SearchContainer from './SearchContainer';
 import useStyles from './styles';
 
-const HeaderSmall = ({ links }) => {
+const links = {
+  about: [{ text: 'Overview', to: '/about' }],
+  join: [{ text: 'Tag Generator', to: '/tag-generator' }],
+  contribute: [
+    { text: 'Index Contributors', to: '/contributors/all' },
+    { text: 'Unaffiliated', to: '/contributors/unaffiliated' },
+    { text: 'Affiliated', to: '/contributors/affiliated' },
+  ],
+  collaborate: [
+    { text: 'Collaborate with Us', to: '/support' },
+    { text: 'How to Do It', to: '/adding-projects-to-the-index' },
+    { text: 'Donate', to: '/donate' },
+    { text: 'Share the CTI', to: '/radicalcollaboration/sharethecti' },
+    { text: 'Volunteer with Us', to: 'https://www.hackforla.org/projects/civic-tech-index', external: true },
+    { text: 'FAQ', to: '/radicalcollaboration/faq' },
+  ],
+};
+
+const HeaderSmall = () => {
   const classes = useStyles();
   const [isBurgerOpen, setIsBurgerOpen] = useState(null);
   const openBurger = () => {
