+++ src/components/Header/index.js
@@ -1,32 +1,19 @@
 import React from 'react';
-import { useTheme } from '@material-ui/core/styles';
-import useMediaQuery from '@material-ui/core/useMediaQuery';
+import Hidden from '@material-ui/core/Hidden';
 import HeaderSmall from './HeaderSmall';
 import HeaderLarge from './HeaderLarge';
 
 const Header = () => {
-  const theme = useTheme();
-  const headerSize = useMediaQuery(theme.breakpoints.up('md'));
-
-  const links = {
-    about: [{ text: 'Overview', to: '/about' }],
-    join: [{ text: 'Tag Generator', to: '/tag-generator' }],
-    contribute: [
-      { text: 'Index Contributors', to: '/contributors/all' },
-      { text: 'Unaffiliated', to: '/contributors/unaffiliated' },
-      { text: 'Affiliated', to: '/contributors/affiliated' },
-    ],
-    collaborate: [
-      { text: 'Collaborate with Us', to: '/support' },
-      { text: 'How to Do It', to: '/adding-projects-to-the-index' },
-      { text: 'Donate', to: '/donate' },
-      { text: 'Share the CTI', to: '/radicalcollaboration/sharethecti' },
-      { text: 'Volunteer with Us', to: 'https://www.hackforla.org/projects/civic-tech-index', external: true },
-      { text: 'FAQ', to: '/radicalcollaboration/faq' },
-    ],
-  };
-
-  return headerSize ? <HeaderLarge links={links} /> : <HeaderSmall links={links} />;
+  return (
+    <>
+      <Hidden mdUp>
+        <HeaderSmall />
+      </Hidden>
+      <Hidden smDown>
+        <HeaderLarge />
+      </Hidden>
+    </>
+  );
 };
 
 export default Header;
